LayoutTests/imported/w3c/web-platform-tests/dom/events/scrolling/wheel-event-transactions-multiple-action-chains-expected.txt


FAIL Two separate webdriver action chains should have different wheel event transactions assert_equals: Wheel event at index `0` did not target the initial element expected Element node <div id="initial"></div> but got Element node <div id="second"></div>
